SG-10799: Ensures we're comparing the current user against the Shotgun user requesting menu actions in a case-insensitive way. (#16)
On Windows 7 we saw issues where a user's credentials would be cached in such a way that a comparison check between the current user in SG Create and the user logged into the web app would fail due to a case sensitivity issue. We're now ensuring that we compare user names in a case insensitive manner.