Modifications to the capability to set a version draft
JIRA: VMR-2289

DESCRIPTION

The previous approach was to create an atomic command, but the way Shotgun
Create works, it's better to have one command that tell the full intention.

Previously, to set a media as a draft version, the user would have to do two
commands and now there's one command that trigger the full logic.